+++ jabsrv/src/main/java/org/jabref/http/server/command/FocusCommand.java
@@ -7,10 +7,6 @@ import jakarta.ws.rs.core.Response;
 
 @JsonTypeName("focus")
 public class FocusCommand extends Command {
-
-    public FocusCommand() {
-    }
-
     @Override
     public Response execute() {
         return execute(new UiCommand.Focus());