+++ jabsrv/src/main/java/org/jabref/http/server/command/SelectEntriesCommand.java
@@ -2,13 +2,8 @@ package org.jabref.http.server.command;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
-import org.jabref.logic.command.CommandSelectionTab;
-import org.jabref.logic.util.io.BackupFileUtil;
-import org.jabref.model.database.BibDatabaseContext;
-import org.jabref.model.entry.BibEntry;
+import org.jabref.logic.UiCommand;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
@@ -17,47 +12,19 @@ import jakarta.ws.rs.core.Response;
 @JsonTypeName("selectentries")
 public class SelectEntriesCommand extends Command {
 
-    @JsonProperty
-    private String libraryId = "";
+    // TODO: This is probably needed for some CAYW functionality to "scope" the citation keys properly
+    // @JsonProperty
+    // private String libraryId = "";
 
-    @JsonProperty
+    @JsonProperty(required = true)
     private List<String> citationKeys = new ArrayList<>();
 
-    @JsonProperty
-    private List<String> entryIds = new ArrayList<>();
-
-    public SelectEntriesCommand() {
-    }
+    // TODO: This is probably needed for some CAYW functionality to use a unique entry id "scope" the citation keys properly
+    // @JsonProperty
+    // private List<String> entryIds = new ArrayList<>();
 
     @Override
     public Response execute() {
-        Optional<CommandSelectionTab> activeTab = getSrvStateManager().getActiveSelectionTabProperty().getValue();
-        if (activeTab.isEmpty()) {
-            return Response.status(Response.Status.BAD_REQUEST)
-                           .entity("This command cannot be executed because no library is opened.")
-                           .build();
-        }
-
-        CommandSelectionTab commandSelectionTab = activeTab.get();
-
-        if (!getLibraryIdFromContext(commandSelectionTab.getBibDatabaseContext()).equals(libraryId)) {
-            return Response.status(Response.Status.BAD_REQUEST)
-                           .entity("This command cannot be executed because the libraryId does not match the active selection tab.")
-                           .build();
-        }
-
-        List<BibEntry> entries = commandSelectionTab.getBibDatabaseContext().getEntries().stream()
-                                                    .filter(entry -> citationKeys.contains(entry.getCitationKey().orElse(null)) || entryIds.contains(entry.getId()))
-                                                    .collect(Collectors.toList());
-
-        commandSelectionTab.clearAndSelect(entries);
-
-        return Response.ok().build();
-    }
-
-    private String getLibraryIdFromContext(BibDatabaseContext bibDatabaseContext) {
-        return bibDatabaseContext.getDatabasePath()
-                                 .map(path -> path.getFileName() + "-" + BackupFileUtil.getUniqueFilePrefix(path))
-                                 .orElse("");
+        return execute(new UiCommand.SelectEntryKeys(citationKeys));
     }
 }